+++ shared/src/api/client/api/documents.ts
@@ -2,7 +2,7 @@ import { Observable, Subscription } from 'rxjs'
 import { createProxyAndHandleRequests } from '../../common/proxy'
 import { ExtDocumentsAPI } from '../../extension/api/documents'
 import { Connection } from '../../protocol/jsonrpc2/connection'
-import { TextDocumentItem } from '../types/textDocument'
+import { ViewComponentData } from '../model'
 import { SubscriptionMap } from './common'
 
 /** @internal */
@@ -11,12 +11,12 @@ export class ClientDocuments {
     private registrations = new SubscriptionMap()
     private proxy: ExtDocumentsAPI
 
-    constructor(connection: Connection, modelTextDocuments: Observable<TextDocumentItem[] | null>) {
+    constructor(connection: Connection, modelViewComponents: Observable<ViewComponentData[] | null>) {
         this.proxy = createProxyAndHandleRequests('documents', connection, this)
 
         this.subscriptions.add(
-            modelTextDocuments.subscribe(docs => {
-                this.proxy.$acceptDocumentData(docs || [])
+            modelViewComponents.subscribe(editors => {
+                this.proxy.$acceptEditorData(editors || [])
             })
         )